# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Easy SKS API
  description: Open-source flashcard study app for the Sportkuestenschifferschein
    (SKS)
  version: 1.0.0
paths:
...
  /navigation-sessions:
    get:
      tags:
      - Navigation
      summary: List Navigation History
      operationId: list_navigation_history_navigation_sessions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NavigationSessionHistoryOut'
                type: array
                title: Response List Navigation History Navigation Sessions Get
    post:
      tags:
      - Navigation
      summary: Start Navigation Session
      operationId: start_navigation_session_navigation_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartNavigationSessionIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavigationSessionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /navigation-sessions/{session_id}:
    get:
      tags:
      - Navigation
      summary: Get Navigation Session
      operationId: get_navigation_session_navigation_sessions__session_id__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavigationSessionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /navigation-sessions/{session_id}/answers/{task_id}:
    put:
      tags:
      - Navigation
      summary: Save Navigation Answer
      operationId: save_navigation_answer_navigation_sessions__session_id__answers__task_id__put
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveNavigationAnswerIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveNavigationAnswerOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /navigation-sessions/{session_id}/submit:
    post:
      tags:
      - Navigation
      summary: Submit Navigation Session
      operationId: submit_navigation_session_navigation_sessions__session_id__submit_post
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavigationResultOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /navigation-sessions/{session_id}/result:
    get:
      tags:
      - Navigation
      summary: Get Navigation Result
      operationId: get_navigation_result_navigation_sessions__session_id__result_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavigationResultOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    NavigationQuestionOut:
      properties:
        task_number:
          type: integer
          title: Task Number
        task_id:
          type: string
          title: Task Id
        points:
          type: integer
          title: Points
        context:
          type: string
          title: Context
        sub_questions:
          items:
            $ref: '#/components/schemas/SubQuestionOut'
          type: array
          title: Sub Questions
        student_answer:
          type: string
          title: Student Answer
        answered_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Answered At
      type: object
      required:
      - task_number
      - task_id
      - points
      - context
      - sub_questions
      - student_answer
      title: NavigationQuestionOut
    NavigationQuestionResultOut:
      properties:
        task_number:
          type: integer
          title: Task Number
        task_id:
          type: string
          title: Task Id
        context:
          type: string
          title: Context
        sub_questions:
          items:
            type: string
          type: array
          title: Sub Questions
        key_answers:
          items:
            type: string
          type: array
          title: Key Answers
        solution_text:
          type: string
          title: Solution Text
        student_answer:
          type: string
          title: Student Answer
        score:
          type: number
          title: Score
        max_score:
          type: number
          title: Max Score
        is_correct:
          type: boolean
          title: Is Correct
        feedback:
          type: string
          title: Feedback
      type: object
      required:
      - task_number
      - task_id
      - context
      - sub_questions
      - key_answers
      - solution_text
      - student_answer
      - score
      - max_score
      - is_correct
      - feedback
      title: NavigationQuestionResultOut
    NavigationResultOut:
      properties:
        session_id:
          type: string
          title: Session Id
        sheet_number:
          type: integer
          title: Sheet Number
        status:
          type: string
          title: Status
        started_at:
          type: string
          title: Started At
        submitted_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Submitted At
        time_limit_minutes:
          type: integer
          title: Time Limit Minutes
        time_over:
          type: boolean
          title: Time Over
        total_score:
          type: number
          title: Total Score
        max_score:
          type: number
          title: Max Score
        passed:
          type: boolean
          title: Passed
        pass_score_threshold:
          type: number
          title: Pass Score Threshold
        questions:
          items:
            $ref: '#/components/schemas/NavigationQuestionResultOut'
          type: array
          title: Questions
      type: object
      required:
      - session_id
      - sheet_number
      - status
      - started_at
      - time_limit_minutes
      - time_over
      - total_score
      - max_score
      - passed
      - pass_score_threshold
      - questions
      title: NavigationResultOut
    NavigationSessionHistoryOut:
      properties:
        session_id:
          type: string
          title: Session Id
        sheet_number:
          type: integer
          title: Sheet Number
        status:
          type: string
          title: Status
        started_at:
          type: string
          title: Started At
        submitted_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Submitted At
        total_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Total Score
        max_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Max Score
        passed:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Passed
        time_over:
          type: boolean
          title: Time Over
      type: object
      required:
      - session_id
      - sheet_number
      - status
      - started_at
      - time_over
      title: NavigationSessionHistoryOut
    NavigationSessionOut:
      properties:
        session_id:
          type: string
          title: Session Id
        sheet_number:
          type: integer
          title: Sheet Number
        status:
          type: string
          title: Status
        started_at:
          type: string
          title: Started At
        submitted_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Submitted At
        deadline_at:
          type: string
          title: Deadline At
        time_limit_minutes:
          type: integer
          title: Time Limit Minutes
        time_remaining_seconds:
          type: integer
          title: Time Remaining Seconds
        time_over:
          type: boolean
          title: Time Over
        task_count:
          type: integer
          title: Task Count
        questions:
          items:
            $ref: '#/components/schemas/NavigationQuestionOut'
          type: array
          title: Questions
      type: object
      required:
      - session_id
      - sheet_number
      - status
      - started_at
      - deadline_at
      - time_limit_minutes
      - time_remaining_seconds
      - time_over
      - task_count
      - questions
      title: NavigationSessionOut
...
    SaveNavigationAnswerIn:
      properties:
        student_answer:
          type: string
          title: Student Answer
          default: ''
      type: object
      title: SaveNavigationAnswerIn
    SaveNavigationAnswerOut:
      properties:
        session_id:
          type: string
          title: Session Id
        task_id:
          type: string
          title: Task Id
        task_number:
          type: integer
          title: Task Number
        student_answer:
          type: string
          title: Student Answer
        answered_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Answered At
      type: object
      required:
      - session_id
      - task_id
      - task_number
      - student_answer
      title: SaveNavigationAnswerOut
...
    StartNavigationSessionIn:
      properties:
        sheet_number:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: Sheet Number
        time_limit_minutes:
          anyOf:
          - type: integer
            maximum: 240.0
            minimum: 1.0
          - type: 'null'
          title: Time Limit Minutes
      type: object
      required:
      - sheet_number
      title: StartNavigationSessionIn
...
    SubQuestionOut:
      properties:
        text:
          type: string
          title: Text
        points:
          type: integer
          title: Points
      type: object
      required:
      - text
      - points
      title: SubQuestionOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...